+++ helpers/buttons_panel_xxx.js
@@ -1,7 +1,9 @@
 ﻿'use strict';
+//07/10/21
 
 include('helpers_xxx_prototypes.js');
 include('helpers_xxx_UI.js');
+include('helpers_xxx_flags.js');
 
 /* 
 	This is the framework to create buttons as new objects with its own properties and tooltips. They can be merged and loaded multiple times
@@ -25,13 +27,6 @@ include('helpers_xxx_UI.js');
 				}
 			}
 */
-
-const ButtonStates = {
-	normal: 0,
-	hover: 1,
-	down: 2,
-	hide: 3
-};
 
 var buttons = {}; // Global list
 var propertiesPrefixes = new Set(); // Global properties names prefixes